requirements.txt
anyio==3.6.2
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpie==3.2.1
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
motor==3.1.1
msgpack==1.0.5
multidict==6.0.4
neovim==0.3.1
orjson==3.8.7
passlib==1.7.4
pycodestyle==2.10.0
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
pynvim==0.4.3
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.3.2
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
